Changed w->is_transient to animation_transient
Changed window is_transient to animation_transient because we can't rely on transient status of the window as sometimes it happens that the flag is removed before unmap animation starts and so window ends up with the wrong animation on close. So now we simply use animation_transient and this is only set when window opens and close animation will check this to be sure it uses the same animation for closing.