Added animation-delta and animation-force-steps options.
This adds animation-delta = 10; configuration option in milliseconds just like for fading which you can use to define the time between animation steps. By default it's set to 10 being the same it was originally set by Daniel.

Also added a new animation-force-steps = true; option which you can use to force the animation steps to not be skipped even if on higher cpu usage so instead of skipping an animation step and going right to the end this will force the script to go trough each and every step even if it's a bit slower instead of just jumping to the end. By default this is disabled so animations behave just as Daniel implemented them and the same way as fading but I prefer it turned on.